feat:display categories number in shopCard

+++ shop-server/src/main/java/fr/fullstack/shopapp/model/Product.java
@@ -40,7 +40,7 @@ public class Product {
     @Column(nullable = false)
     @PositiveOrZero(message = "Price must be positive")
     @NotNull(message = "Price may not be null")
-    private float price;
+    private Long price; // stocker les centimes
 
     @ManyToOne
     private Shop shop;
@@ -57,7 +57,7 @@ public class Product {
         return localizedProduct;
     }
 
-    public float getPrice() {
+    public Long getPrice() {
         return price;
     }
 
@@ -77,7 +77,7 @@ public class Product {
         this.localizedProduct = localizedProduct;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
